requirements.txt
Django==5.2.1
Pillow
psycopg2
python-dotenv
django-crum
